changed how mazes are stored

+++ src/lib.rs
@@ -3,8 +3,8 @@ pub mod maze;
 pub mod consts {
     use sfml::graphics::Color;
 
-    pub const MAZE_WIDTH: usize = 64;
-    pub const MAZE_HEIGHT: usize = 64;
+    pub const MAZE_WIDTH: u16 = 64;
+    pub const MAZE_HEIGHT: u16 = 64;
     pub const CELL_SIZE: usize = 8;
     pub const WALL_WIDTH: usize = 1;
